requirements.txt
PyJWT == 2.10.1
Flask == 3.1.0